Integrated source code from branch into respective master folder locations
