+++ src/data_collection.py
@@ -1,5 +1,55 @@
 """ Collects commit data for contributors of the master branch of a repo. """
 from pydriller import RepositoryMining
+from github import Github
+import json_handler
+
+
+def authenticate_repository(user_token, repository_name):
+    """Authenticate the Github repository using provided credentials."""
+
+    # Credentials for PyGithub functions and methods
+    ghub = Github(user_token)
+    repository = ghub.get_repo(repository_name)
+
+    return repository
+
+
+# Written as a temporary pass-through in case this variable is converted to a global
+# variable, in which case that process would occur here. Pass-through will be eliminated
+# during refactoring.
+def initialize_contributor_data(file_path):
+    """Load a dictionary based upon the given .json file."""
+
+    contributor_data = json_handler.get_dict_from_json_file(file_path)
+
+    return contributor_data
+
+
+def retrieve_issue_data(repository, state, contributor_data):
+    """Retrieve a contributor's involvement based upon issues and pull request threads."""
+
+    issues = repository.get_issues(state=state)
+
+    for issue in issues:
+        for comment in issue.get_comments():
+            if comment.user.login in contributor_data.keys():
+                if issue.pull_request is None:
+                    contributor_data[comment.user.login]["issues_commented"].append(
+                        issue.number
+                    )
+                else:
+                    contributor_data[comment.user.login][
+                        "pull_requests_commented"
+                    ].append(issue.number)
+        if issue.user.login in contributor_data.keys():
+            if issue.pull_request is None:
+                contributor_data[issue.user.login]["issues_opened"].append(issue.number)
+            else:
+                contributor_data[issue.user.login]["pull_requests_opened"].append(
+                    issue.number
+                )
+
+    return contributor_data
 
 
 def collect_commits_hash(repo):
